Refactor: client-utils and response-utils
